pyproject.toml
[tool.poetry]
name = "pseudonym-se"
version = "0.0.1"
description = ""
authors = ["Jeremy Herbst"]
license = "MIT"
readme = "README.md"
packages = [{include = "pseudonym_se"}]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.83.0"
redis = "^4.3.4"

[tool.poetry.group.charm.dependencies]
charm-crypto = {path = "../charm"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
